+++ src/VecSim/utils/vec_utils.h
@@ -10,6 +10,7 @@
 #include "VecSim/vec_sim_common.h"
 #include <VecSim/query_results.h>
 #include <utility>
+#include <cassert>
 #include <cmath> //sqrt
 
 template <typename dist_t>
@@ -56,9 +57,24 @@ public:
     static const char *BATCH_SIZE_STRING;
 };
 
-int cmpVecSimQueryResultById(const VecSimQueryResult *res1, const VecSimQueryResult *res2);
+inline int cmpVecSimQueryResultById(const VecSimQueryResult *res1, const VecSimQueryResult *res2) {
+    return (int)(VecSimQueryResult_GetId(res1) - VecSimQueryResult_GetId(res2));
+}
 
-int cmpVecSimQueryResultByScore(const VecSimQueryResult *res1, const VecSimQueryResult *res2);
+inline int cmpVecSimQueryResultByScore(const VecSimQueryResult *res1,
+                                       const VecSimQueryResult *res2) {
+    assert(!std::isnan(VecSimQueryResult_GetScore(res1)) &&
+           !std::isnan(VecSimQueryResult_GetScore(res2)));
+    // Compare doubles
+    return (VecSimQueryResult_GetScore(res1) - VecSimQueryResult_GetScore(res2)) >= 0.0 ? 1 : -1;
+}
+
+inline int cmpVecSimQueryResultByScoreThenId(const VecSimQueryResult *res1,
+                                             const VecSimQueryResult *res2) {
+    return (VecSimQueryResult_GetScore(res1) != VecSimQueryResult_GetScore(res2))
+               ? cmpVecSimQueryResultByScore(res1, res2)
+               : cmpVecSimQueryResultById(res1, res2);
+}
 
 void sort_results_by_id(VecSimQueryResult_List results);
